+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - eslint-plugin-ember-suave
  - ember-cli-anybar
  - ember-cli-dependency-lint
+ - ember-native-dom-helpers
 
 ### Updated
   - Code style updates for ember-suave